Apply the same draining technique to the partial results
This somehow significantly improves the tracking - presumably we're just occassionally getting stuck on a few words of mismatch that throw the whole transcript out of alignment, and this approach enables us to reconnect the alignment. Because the partial state is only temporary, it doesn't do any lasting damage to the synchronisation, which to the user presents as sometimes the prompter gets a little bit ahead of them but no bother, because they can still see what they need to read out.